Fix infinite loop waiting for events
Fix an unknown event occurs the loop could potentially go on
forever. Track wallclock time to ensure an absolute upper
bound on waiting for events

+++ src/backend/entangle-camera.h
@@ -91,7 +91,7 @@ gboolean entangle_camera_delete_file(EntangleCamera *cam,
 
 gboolean entangle_camera_event_flush(EntangleCamera *cam);
 gboolean entangle_camera_event_wait(EntangleCamera *cam,
-                                int waitms);
+                                    guint64 waitms);
 
 gboolean entangle_camera_get_has_capture(EntangleCamera *cam);
 gboolean entangle_camera_get_has_preview(EntangleCamera *cam);
